fixed endpoints

+++ dotnet/FinkiRasporedi/Repository/Impl/ScheduleRepository.cs
@@ -51,38 +51,55 @@ namespace FinkiRasporedi.Repository
             return schedule;
         }
 
+        public async Task<Schedule?> AddDuplicateLectureAsync(int id, int original_lectureId)
+        {
+            Lecture lecture = await _lectureRepository.GetByIdAsync(original_lectureId);
+            if (lecture.Type == 0)
+            {
+
+
+                Schedule schedule = await GetByIdAsync(id);
+                Lecture duplicateLecture = new()
+                {
+                    Name = lecture.Name,
+                    OriginalLecture = lecture,
+                    Day = lecture.Day,
+                    TimeFrom = lecture.TimeFrom,
+                    TimeTo = lecture.TimeTo,
+                    Professor = lecture.Professor,
+                    Course = lecture.Course,
+                    Room = lecture.Room,
+                    Type = 1
+                };
+                Lecture newLecture = await _lectureRepository.AddAsync(duplicateLecture);
+                schedule.Lectures.Add(newLecture);
+                schedule.Lectures.Remove(lecture);
+                _context.Entry(schedule).State = EntityState.Modified;
+                await _context.SaveChangesAsync();
+                return schedule;
+            }
+            return null;
+        }
 
         public async Task<Schedule> RemoveLectureAsync(int id, int lectureId)
         {
             Lecture lecture = await _lectureRepository.GetByIdAsync(lectureId);
             Schedule schedule = await GetByIdAsync(id);
             schedule.Lectures.Remove(lecture);
+            if (lecture.OriginalLecture != null)
+            {
+                schedule.Lectures.Add(lecture.OriginalLecture);
+            }
+
+            if (lecture.Type == 1 || lecture.Type == 2)
+            {
+                await _lectureRepository.DeleteAsync(lectureId);
+            }
             _context.Entry(schedule).State = EntityState.Modified;
             await _context.SaveChangesAsync();
             return schedule;
         }
 
-
-        public async Task<Schedule> AddCustomLectureAsync(int id, int lectureId, string name)
-        {
-            Lecture oldLecture = await _lectureRepository.GetByIdAsync(lectureId);
-            Schedule schedule = await GetByIdAsync(id);
-            var customLecture = await _lectureRepository.AddCustomAsync(oldLecture, name);
-            schedule.Lectures.Add(customLecture);
-            _context.Entry(schedule).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return schedule;
-        }
-        public async Task<Schedule> RemoveCustomLectureAsync(int id, int lectureId, string name)
-        {
-            Lecture customLecture = await _lectureRepository.GetByIdAsync(lectureId);
-            Schedule schedule = await GetByIdAsync(id);
-            schedule.Lectures.Remove(customLecture);
-            await _lectureRepository.DeleteAsync(lectureId);
-            _context.Entry(schedule).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return schedule;
-        }
         public async Task<Schedule> DeleteAsync(int id)
         {
             var schedule = await GetByIdAsync(id);
